requirements.txt
Flask==3.0.2
PyYAML==6.0.1
gunicorn==21.2.0